+++ packages/luce/example/luce_example.dart
@@ -30,14 +30,17 @@ class CounterComponent extends Component {
   @override
   Widget build(BuildContext context) {
     context.rebuildOn(counter.changes);
-    return Div(<Widget>[
-      const Txt('Your Luce app is running.'),
-      const Br(),
-      Flag(
-        className: 'high',
-        when: counter.value > 7,
-        child: Txt('Counter value is ${counter.value}'),
-      ),
-    ]);
+    return MouseEvents(
+      onClick: (MouseEvent e) => counter.up(),
+      child: Div(<Widget>[
+        const Txt('Your Luce app is running.'),
+        const Br(),
+        Flag(
+          className: 'high',
+          when: counter.value > 7,
+          child: Txt('Counter value is ${counter.value}'),
+        ),
+      ]),
+    );
   }
 }