+++ packages/luce/lib/luce.dart
@@ -1,7 +1,9 @@
+export 'dart:async';
 export 'dart:html';
 
 export 'src/core/attributes.dart';
 export 'src/core/components.dart';
+export 'src/core/events.dart';
 export 'src/core/state.dart';
 export 'src/core/tags.dart';
 export 'src/core/txt.dart';