+++ packages/luce/lib/src/core/components.dart
@@ -1,3 +1,4 @@
+import 'dart:html';
 import 'state.dart';
 import 'vdom.dart';
 
@@ -7,21 +8,55 @@ abstract class Component extends Widget {
   Widget build(BuildContext context);
 
   @override
-  VNode createVNode(BuildRoot parent) => VComponent(this, parent);
+  VComponent createVNode(BuildRoot parent) => VComponent(this, parent);
 }
 
-class VComponent extends VSingleChildNode<Component> implements BuildContext {
-  VComponent(Component widget, BuildRoot parent) : super(widget, parent) {
-    child = widget.build(this).createVNode(parent);
-    node = child.node;
+class VComponent extends VNode implements BuildContext {
+  VComponent(this.widget, BuildRoot parent) : super(parent) {
+    child = widget.build(this).createVNode(this);
+    _childNode = child.node;
   }
 
+  @override
+  Component widget;
+  @override
+  Element get element => _childElement ??= child.element;
+  @override
+  Node get node => _childElement ?? _childNode;
+
+  Element _childElement;
+  Node _childNode;
+  VNode child;
   final List<RemoveListener> _removeListeners = <RemoveListener>[];
 
   @override
-  VNode updateAndReturnChild(Component newWidget) {
-    removeListeners();
-    return child.update(newWidget.build(this));
+  VNode update(Widget newWidget) {
+    if (!hasDirtyChild && newWidget == widget) {
+      return this;
+    }
+    hasDirtyChild = false;
+    if (newWidget is Component) {
+      if (!isDirty && newWidget == widget) {
+        child.update(child.widget);
+        _refreshChildNode();
+      } else {
+        isDirty = false;
+        widget = newWidget;
+        removeListeners();
+        child = child.update(newWidget.build(this))..parent = this;
+        _refreshChildNode();
+      }
+      return this;
+    } else {
+      return newWidget.createVNode(invalidate());
+    }
+  }
+
+  void _refreshChildNode() {
+    if (_childNode != child.node) {
+      _childNode = child.node;
+      _childElement = null;
+    }
   }
 
   @override
@@ -39,6 +74,13 @@ class VComponent extends VSingleChildNode<Component> implements BuildContext {
   @override
   BuildRoot invalidate() {
     removeListeners();
+    widget = null;
+    _childNode = null;
+    _childElement = null;
+    if (child != null) {
+      child.invalidate();
+      child = null;
+    }
     return super.invalidate();
   }
 }